+++ CHANGELOG.md
@@ -3,7 +3,15 @@ title: Changelog - Alfresco Solr Query Analytics
 
 # Alfresco Solr Query Analytics Changelog
 
-## Unreleased
+## v0.0.3 - 10-03-2022
+
+### Added
+* [ALFREDOPS-814] Disable logging for search check[#20]
+
+### Changed
+* [ALFREDOPS-813] Refactor to SolrDebugQueryHTTPClient[#19]
+
+## v0.0.2 - 30-09-2021
 
 ### Added
 * [ETHIASSLA-422] Custom appender for log rotation
@@ -23,5 +31,7 @@ Fix publishing via github actions.
 [#8]: https://github.com/xenit-eu/alfresco-solr-debug/pull/8
 [#12]: https://github.com/xenit-eu/alfresco-solr-debug/pull/12
 [#13]: https://github.com/xenit-eu/alfresco-solr-debug/pull/13
+[#19]: https://github.com/xenit-eu/alfresco-solr-debug/pull/19
+[#20]: https://github.com/xenit-eu/alfresco-solr-debug/pull/20